lyedu-api-python/requirements.txt
# LyEdu API - Python 版本 (Python 3.10+，含 3.14)
# Python 3.14 需使用较新 pydantic 才有预编译 wheel，故未严格锁版
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
pymysql>=1.1.0
PyJWT>=2.10.0
bcrypt>=4.0.0
passlib[bcrypt]>=1.7.4
python-dotenv>=1.0.0
pydantic>=2.10.0
pydantic-settings>=2.6.0
